Avoid passing (another) copy of the full data to workers
bpmapply was passing another copy of the full data. In
this case, directly as an extra argument to bplapply.

The arguments to be iterated on are now transposed,
so instead of having ddd[[i]][[j]] being the value of the
i-th argument in the j-th iteration, we build a list of
the form ddd[[j]][[i]]. Then bplapply can directly
iterate on ddd, passing one element of ddd to each worker,
instead of passing the whole ddd list of lists and the
corresponding index.

This approach has a drawback when the arguments to be
iterated on are not lists, but vectors. In that case,
the transposition has some overhead, but I would argue
that it would make more sense to use something like
bpvec or bpmvec if it existed.