bpmapply: Use bpsharememory() to avoid extra copies of all data
bpmapply receives several arguments to iterate on, each of them.

This ends up being something like:

ddd <- list(
  arg1 = list(arg1_iter1, arg1_iter2, arg1_iter3),
  arg2 = list(arg2_iter1, arg2_iter2, arg2_iter3)
)

The implementation before this commit was passing ddd to all
workers as well as the iteration index, and each worker would
take the corresponding slice ddd$arg1[[i]] and ddd$arg2[[i]].

For Serial and MultiCore backends, where workers share memory,
this is very efficient. However for Snow backends, where workers
do not share memory, the whole ddd needs to be serialized, copied
to each worker and deserialized, which is very inneficient.

In the Snow scenario, it is far more efficient to let the main
process "transpose" the `ddd` list so it becomes:

ddd2 <- list(
  iter1 = list(arg1_iter1, arg2_iter1),
  iter2 = list(arg1_iter2, arg2_iter2)
)

Then only pass the corresponding iteration to each worker,
reducing the amount of serialization/deserialization needed
and the memory footprint significantly.

For this to be efficient in all backends, we need to know
if the specific backend has workers that share memory or not.

This commit uses the newly introduced bpsharememory() generic
to determine that and choose the most efficient bpmapply
strategy.